django_logzio_otel_trace/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Deprecated==1.2.14
Django==4.2.13
googleapis-common-protos==1.63.0
grpcio==1.64.0
idna==3.7
importlib_metadata==7.1.0
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-django==0.46b0
opentelemetry-instrumentation-wsgi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
protobuf==4.25.3
requests==2.32.3
setuptools==70.0.0
sqlparse==0.5.0
typing_extensions==4.12.1
urllib3==2.2.1
wrapt==1.16.0
zipp==3.19.1
